requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.2
Flask-Login==0.6.3
Flask-WTF==1.1.1
python-dotenv==1.0.0
SQLAlchemy==1.4.50
Werkzeug==2.3.7
PyJWT==2.8.0
bcrypt==4.1.2
python-dateutil==2.8.2
Flask-Cors==4.0.0
numpy==1.24.3
